Clean up some code and fix bugs
Code quality in the parser is still questionable at best,
but I reorganized some parts to improve readability. This
commit also fixes a bug with `pop_substr()` and `popl_substr()`
where a segfault would occur if there was not a trailing
delimiter. No idea how I originally missed that, but it's fixed
now. `getline()` also no longer returns a string ending with
a newline.